Rename NewExceptionDataFactory
